feat : 딜러 구현 - 추상화를 활용
