+++ README.md
@@ -20,3 +20,13 @@
     [x] Ace 판단
 [x] 게임의 흐름 제어  
 [x] Request Response  
+
+
+step3
+[] 딜러 구현
+    [x] 추상화 or 인터페이스의 조합을 활용
+    [] 딜러의 추가 기능 구현
+        [] 16이하 일 때 카드를 더 받는 것
+        [] 21초과시 버스트 상태와 함께 모든 유저 승리
+[] 스코어 결과에 따른 승패 상태 추가
+[] 피드백 반영
